Update notification list tests
